Add assimilation engine
Based on github.com/arodland/giroviz, based on github.com/AF7TI/giroviz
(but just does the interpolation and data merging parts, not the map
rendering)

Uses a 3D Gaussian Process model to smoothly warp the IRI-model data
based on observed (or predicted) ionosonde data, so that the values near
station locations is consistent with those observations, but values far
from station locations makes use of the expected variations from the
IRI-model, instead of being generally blank.

foF2 and hmF2 are calculated directly; MUF(3000) is calculated by
creating an assimilated map of M(3000) (i.e. MUF(3000) / foF2) and
multiplying it by the assimilated map of foF2. Currently the advantage
of this approach is minimal, but given some improvements to the pred
engine it should allow stations that contribute foF2 but not MUF(3000)
measurements to contribute to the assimilated MUF(3000) map.