Add scheduler
This is the thing that schedules runs of everything on a periodic basis.

Every run begins with an eSSN calculation (the ID of the ESSN job
becomes the "run ID") and a pred calculation for one or more target
times.

The completion of an eSSN job allows starting irimap jobs with that eSSN
value.

When pred and irimap are both complete, we can run an assimilate job
with those products.

When assimilate is complete, we can run several rendersvg jobs.

When all rendersvg jobs are complete, we can run a renderhtml job, which
sets all of the maps as active.