noaa-loader: fix loading converted IONFM files
not sure if this ever worked... for some reason POSIX::gmtime doesn't
work when called with no argument. Not sure why I was importing gmtime
from POSIX either, since it's documented to be exactly the same as the
core version (but then again it's not!). In any case stop importing it.