raytrace: replace interpolator
use a numba-compiled bilinear interpolator for the ionospheric
parameters instead of scipy.interpolate.RectBivariateSpline.

It's faster and the precision loss is tiny.

I tried doing the same thing for assimilate: it works, but the
interpolation is a much smaller portion of the work there, so it doesn't
make any significant difference.