scheduler: add job expiration
Don't let anything run so late that it would overlap the next job.

It might seem like a decent idea to run the computation stuff even if
it's behind, and just skip the 'renderhtml' job which points 'current'
to that run, but essn, pred, and assimilate use *all* available
information, they don't ignore data that's newer than the run_id
timestamp. So any data produced significantly late is scientifically
useless, and redundant. We should just make the effort to catch up with
the next run.