New experiment: train separate time kernel per metric
Note that since there's only one 'cs' field in the prediction table, we
can only have one uncertainty per record, and so pred picks mufd if it
exists, else fof2, else hmf2. Obviously this isn't great and we need to
overhaul the prediction schema (there are other reasons to do so, for
that matter, like passing the stdev directly, and passing the log
directly for assimilate to use instead of making it re-subtract from an
interpolated IRI). But not today.