Add -D 1 to minion workers
minion has postgres triggers that are supposed to notify the workers
when there's new work, but they don't quite do it when a parent job
unblocks a child job, so instead we end up waiting for the 5-second poll
to timeout. Reduce that poll to 1 second to get everything completing a
bit faster.