giro-loader: go one more period into the past
We've been getting a bunch of 503s lately. The current code looks at the
last successful reception of *any* record from GIRO to decide how far
back to request. That handles a total GIRO outage or the giro-loader
being down for a period of time, but not a sporadic failure of some
stations to fetch. Give the failures a second chance on the next go.