api: fix several endpoints to use get_latest_run
caching was wrong for all of these when using the implicit "latest"
run/timestamp. Instead of having them do their own query, just call
get_latest_run if the params aren't present, and then use the run/ts we
get from that in the cache key, which is much much better (and enables
sharing cache with the explicit run/ts calls that the backend uses).

In addition, add caching to get_latest_run to make these new calls more
efficient.