api: fix arrow streaming to actually stream
c13156ced98bf1764482475110947291d6be7023 made it *work*, but it had the
side effect of breaking the query.yield_per setting, meaning that it
would fetch the whole resultset from Pg into RAM before sending anything
to the client, which is a waste.

This is an ugly fix but it works.