ipe: fix trying to fetch runs in the future
we would start with the 6-hour interval previous to the target time and
work backwards, which meant that if the target time was 24h from now it
would start by looking for a run generated 18 hours in the future! Not
very productive. Reduce the 404s by only considering times in the past.