create a new RoPE-style param encoder for conditioned diffusion
not sure how necessary this is, but it kind of can't hurt. And what
definitely can't hurt is that the outputs are now scaled to (-1,1) via a
tanh output, instead of being strictly positive via a ReLU output,
meaning that the zero vector will now be somewhere in the middle of the
embedding space instead of way out in la-la land.